Refactor larpix.py into submodules (#187)
* Refactor larpix.py into submodules

* Update docs

* Optimize functions called in dataserver_message_decode (#185)

* Optimize functions called in dataserver_message_decode

Re: #179

* Modify io protocol, data packets, and message format encoding to remove Key generation from reads

* Fix bug in zmq_io decode method

* Fix miso map issue in MultiZMQ_IO, fix bug in MultiZMQ_IO cleanup

* Fix bug in MultiZMQ_IO miso_map

* Adapt tests to use Configuration.num_channels

* Remove import warning and add ability to import from larpix directly

* Revamp import paths, update README + tutorial tests

* Fix failed test
