+++ larpix/bitarrayhelper.py
@@ -5,13 +5,13 @@ A module with convenience functions for bitarray.
 from bitarray import bitarray
 
 def fromuint(val, nbits):
-    if isinstance(val, bitarray):
+    try:
+        if isinstance(nbits, slice):
+            nbits = abs(nbits.stop - nbits.start)
+        string = bin(val)[2:].zfill(nbits)
+        return bitarray(string)
+    except TypeError:
         return val
-    if isinstance(nbits, slice):
-        nbits = abs(nbits.stop - nbits.start)
-    bin_string = bin(val)[2:]
-    padding = '0' * (nbits - len(bin_string))
-    return bitarray(padding + bin_string)
 
 def touint(bits):
     bin_string = bits.to01()
