+++ larpix/format/message_format.py
@@ -11,7 +11,7 @@ import struct
 
 from larpix.larpix import Packet, TimestampPacket
 
-def dataserver_message_encode(packets, key_parser=None, version=(1,0)):
+def dataserver_message_encode(packets, version=(1,0)):
     r'''
     Convert a list of packets to larpix dataserver messages. DAQ board messages are formatted using 8-byte words with the first word being a header word describing the interpretation of other words in the message. These messages are formatted as follows
 
@@ -39,22 +39,17 @@ def dataserver_message_encode(packets, key_parser=None, version=(1,0)):
          - byte[8:14] = 7-byte Unix timestamp
          - byte[15] is unused
 
-    A key parser should be provided to extract the ``'io_chain'`` from the packet chip key. If none is provided, io_chain
-    will be 0 for all packets. E.g.::
+    Example usage::
 
         from larpix.larpix import Packet, Key
-        def ex_key_parser(key):
-            return dict(io_chain=key.io_channel)
         packet = Packet()
         packet.chip_key = Key('1-1-1')
-        msgs = datserver_message_encode([packet], key_parser=ex_key_parser)
+        msgs = datserver_message_encode([packet])
         msgs[0] # b'\x01\x00D\x01\x00\x00\x00\x00\x04\x00\x00\x00\x00\x00\x00\x00'
         msgs[0][:8] # header
         msgs[0][8:] # data words
 
     :param packets: list of ``larpix.Packet`` objects
-
-    :param key_parser: optional, a method that takes a ``larpix.Key`` object and returns a dict with ``'io_chain'``
 
     :param version: optional, encode a message in specified version format, ``tuple`` of major, minor numbers
 
@@ -70,8 +65,10 @@ def dataserver_message_encode(packets, key_parser=None, version=(1,0)):
             header = [0]*len(data_header_fmt)
             header[0:2] = version[0:2]
             header[2] = b'D'
-            if key_parser:
-                header[3] = key_parser(packet.chip_key)['io_chain']
+            if packet.io_channel:
+                header[3] = packet.io_channel
+            else:
+                raise ValueError('all packets must have a declared io_channel')
             msg = struct.pack(data_header_fmt, *header)
             msg += packet.bytes() + struct.pack('B',0)
         elif isinstance(packet, TimestampPacket):
@@ -83,28 +80,15 @@ def dataserver_message_encode(packets, key_parser=None, version=(1,0)):
         msgs += [msg]
     return msgs
 
-def dataserver_message_decode(msgs, key_generator=None, version=(1,0), **kwargs):
+def dataserver_message_decode(msgs, version=(1,0), **kwargs):
     r'''
-    Convert a list of larpix data server messages into packets. A key generator
-    should be provided if packets are to be used with an ``larpix.io.IO``
-    object. The data server messages provide a ``chip_id`` and ``io_chain`` for
-    keys. Additional keyword arguments can be passed along to the key generator. E.g.::
-
-        from larpix.larpix import Key
-        def ex_key_gen(chip_id, io_chain, io_group):
-            return Key(Key.key_format.format(
-                chip_id=chip_id,
-                io_channel=io_chain,
-                io_group=io_group
-            ))
+    Convert a list of larpix data server messages into packets. Additional packet meta data can be passed along via kwargs E.g.::
 
         msg = b'\x01\x00D\x01\x00\x00\x00\x00\x04\x00\x00\x00\x00\x00\x00'
-        packets = dataserver_message_decode([msg], key_generator=ex_key_gen, io_group=1)
+        packets = dataserver_message_decode([msg], io_group=1)
         packets[0] # Packet(b'\x04\x00\x00\x00\x00\x00\x00'), key of '1-1-1'
 
     :param msgs: list of bytestream messages each starting with a single 8-byte header word, followed by N 8-byte data words
-
-    :param key_generator: optional, a method that takes ``chip_id`` and ``io_chain`` as arguments and returns a ``larpix.Key`` object
 
     :param version: optional, message version to validate against, ``tuple`` of major, minor version numbers
 
@@ -120,6 +104,9 @@ def dataserver_message_decode(msgs, key_generator=None, version=(1,0), **kwargs)
         if msg_type == b'T':
             timestamp = struct.unpack('L',msg[8:15] + b'\x00')[0] # only use 7-bytes
             packets.append(TimestampPacket(timestamp=timestamp))
+            if kwargs:
+                for key,value in kwargs.items():
+                    setattr(packets[-1], key, value)
         elif msg_type == b'D':
             io_chain = struct.unpack('B',msg[3:4])[0]
             payload = msg[8:]
@@ -127,8 +114,10 @@ def dataserver_message_decode(msgs, key_generator=None, version=(1,0), **kwargs)
                 for start_index in range(0, len(payload), 8):
                     packet_bytes = payload[start_index:start_index+7]
                     packets.append(Packet(packet_bytes))
-                    if key_generator:
-                        packets[-1].chip_key = key_generator(chip_id=packets[-1].chipid, io_chain=io_chain, **kwargs)
+                    packets[-1].io_channel = io_chain
+                    if kwargs:
+                        for key,value in kwargs.items():
+                            setattr(packets[-1], key, value)
         elif msg_type == b'H':
             print('Heartbeat message: {}'.format(msg[3:]))
     return packets
