+++ larpix/logger/stdout_logger.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 import time
 
 from larpix.logger import Logger
-from larpix.larpix import Packet
+from larpix import Packet
 
 class StdoutLogger(Logger):
     '''
